ai_service/requirements.txt
# Service IA pour CAMEG-CHAIN
fastapi==0.115.6
uvicorn[standard]==0.32.1
pydantic==2.10.3
python-dotenv==1.0.1
httpx==0.28.1
requests==2.32.3

# Machine Learning (optionnel pour l'instant)
# scikit-learn==1.5.2
# pandas==2.2.3
# numpy==1.26.4

# Monitoring
structlog==24.4.0